WiggleR/Code/Api/pyproject.toml
[project]
name = "wiggler_studiorabota"
version = "0.0.46"
authors = [
  { name="Vincent Kranendonk" },
]
description = "WiggleR - Desktop research unit for making controlled experiments"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
  'fastapi[all]>=0.68.0,<0.69.0',
  'python-crontab>=3.0.0,<4.0.0',
  'rpi_ws281x',
  'adafruit-circuitpython-neopixel',
  'pillow'
]

[project.scripts]
wiggler = "wiggler.cli:main"

[project.urls]
"Homepage" = "https://github.com/studiorabota/wiggle-bin"